Refactor: delete de unused articles
